LayoutTests/http/tests/misc/iframe-beforeunload-dialog-not-matching-ancestor-securityorigin.html

<script>

if (window.testRunner) {
	testRunner.dumpAsText();
	testRunner.waitUntilDone();
}

window.onload = function()
{
	window.setTimeout("window.location.href = 'resources/notify-done.html';", 0);
}

</script>
<body>
When viewed using the host 127.0.0.1, this page has an iframe whose security origin is different from the main frame.<br>
The iframe has a beforeunload handler, and that handler should not result in the display of the page dismissal dialog since the security origins differ.<br>
<iframe src="http://localhost:8000/misc/resources/iframe-with-beforeunload.html"></iframe><br>
</body>
