LayoutTests/http/tests/misc/iframe-beforeunload-dialog-not-matching-ancestor-securityorigin-expected.txt

CONSOLE MESSAGE: Blocked attempt to show beforeunload confirmation dialog on behalf of a frame with different security origin. Protocols, domains, and ports must match.
